Feat : 동아리장 신청서 리스트 확인 로직 위치 변경
