+++ src/main/java/com/greedy/mokkoji/api/admin/controller/AdminController.java
@@ -1,0 +1,33 @@
+package com.greedy.mokkoji.api.admin.controller;
+
+import com.greedy.mokkoji.api.admin.dto.response.GetClubMasterApplicationsResponse;
+import com.greedy.mokkoji.api.admin.service.AdminService;
+import com.greedy.mokkoji.api.auth.controller.argumentResolver.AuthCredential;
+import com.greedy.mokkoji.api.auth.controller.argumentResolver.Authentication;
+import com.greedy.mokkoji.common.response.APISuccessResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/admin")
+public class AdminController {
+
+    private final AdminService adminService;
+
+    @GetMapping("club-master-applications")
+    public ResponseEntity<APISuccessResponse<List<GetClubMasterApplicationsResponse>>> getMyClubMasterApplications(
+            @Authentication final AuthCredential authCredential
+    ) {
+        return APISuccessResponse.of(
+                HttpStatus.OK,
+                adminService.getClubMasterApplications(authCredential.authRole(), authCredential.userId())
+        );
+    }
+}
