+++ src/main/java/com/greedy/mokkoji/api/clubMaster/service/ClubMasterService.java
@@ -1,7 +1,6 @@
-package com.greedy.mokkoji.api.clubMasterApplication.service;
+package com.greedy.mokkoji.api.clubMaster.service;
 
-import com.greedy.mokkoji.api.clubMasterApplication.dto.response.GetClubMasterApplicationsResponse;
-import com.greedy.mokkoji.api.clubMasterApplication.dto.response.GetMyClubMasterApplicationsResponse;
+import com.greedy.mokkoji.api.clubMaster.dto.response.GetMyClubMasterApplicationsResponse;
 import com.greedy.mokkoji.common.exception.MokkojiException;
 import com.greedy.mokkoji.db.club.entity.Club;
 import com.greedy.mokkoji.db.club.repository.ClubRepository;
@@ -11,7 +10,6 @@ import com.greedy.mokkoji.db.university.entity.University;
 import com.greedy.mokkoji.db.university.repository.UniversityRepository;
 import com.greedy.mokkoji.db.user.entity.User;
 import com.greedy.mokkoji.db.user.repository.UserRepository;
-import com.greedy.mokkoji.enums.auth.AuthRole;
 import com.greedy.mokkoji.enums.message.FailMessage;
 import com.greedy.mokkoji.enums.university.UniversityCode;
 import lombok.RequiredArgsConstructor;
@@ -23,7 +21,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class ClubMasterApplicationService {
+public class ClubMasterService {
 
     private final ClubMasterApplicationRepository clubMasterApplicationRepository;
     private final UniversityRepository universityRepository;
@@ -74,32 +72,6 @@ public class ClubMasterApplicationService {
                 .toList();
     }
 
-    @Transactional
-    public List<GetClubMasterApplicationsResponse> getClubMasterApplications(
-            final AuthRole authRole,
-            final Long userId
-    ) {
-        validateAdminRole(authRole);
-
-        User user = getUserOrThrow(userId);
-        Long universityId = user.getUniversity().getId();
-
-        List<ClubMasterApplication> applications = clubMasterApplicationRepository.findByUniversityIdOrderByCreatedAtAsc(universityId);
-
-        return applications.stream()
-                .map(application -> new GetClubMasterApplicationsResponse(
-                        application.getId(),
-                        application.getUniversity().getName(),
-                        application.getClub().getName(),
-                        application.getUserName(),
-                        application.getStatus(),
-                        application.getRejectReason(),
-                        application.getCreatedAt(),
-                        application.getUpdatedAt()
-                ))
-                .toList();
-    }
-
     private University getUniversityOrThrow(final UniversityCode universityCode) {
         return universityRepository.findByUniversityCode(universityCode)
                 .orElseThrow(() -> new MokkojiException(FailMessage.NOT_FOUND_UNIVERSITY));
@@ -113,11 +85,5 @@ public class ClubMasterApplicationService {
     private User getUserOrThrow(final Long userId) {
         return userRepository.findById(userId)
                 .orElseThrow(() -> new MokkojiException(FailMessage.NOT_FOUND_USER));
-    }
-
-    private void validateAdminRole(AuthRole authRole) {
-        if (!AuthRole.ADMIN.equals(authRole)) {
-            throw new MokkojiException(FailMessage.FORBIDDEN_MANAGE_UNIVERSITY_CLUB);
-        }
     }
 }
